Split RadialPotential to a separate file
